Different file_exists for windows and other
