+++ pkg/fileutil/files.go
@@ -3,15 +3,12 @@ package fileutil
 import (
 	"bytes"
 	"encoding/hex"
-	"errors"
 	"fmt"
-	"golang.org/x/sys/windows"
 	"io"
 	"math/rand"
 	"os"
 	"path/filepath"
 	"strings"
-	"syscall"
 	"text/template"
 
 	"runtime"
@@ -124,19 +121,6 @@ func CopyDir(src string, dst string) error {
 	}
 
 	return nil
-}
-
-// FileExists checks a file's existence
-func FileExists(name string) bool {
-	if _, err := os.Stat(name); err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
-		if errors.Is(err, syscall.Errno(windows.ERROR_INVALID_NAME)) {
-			return false
-		}
-	}
-	return true
 }
 
 // IsDirectory returns true if path is a dir, false on error or not directory
